Extract Language Client logic to service
